# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.SysApi:
    properties:
      apiGroup:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      method:
        type: string
      path:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.SysDictionaryDetail:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      sort:
        type: integer
      status:
        type: boolean
      sysDictionaryID:
        type: integer
      updatedAt:
        type: string
      value:
        type: integer
    type: object
...
  request.GetById:
    properties:
      id:
        type: number
    type: object
...
  request.SearchApiParams:
    properties:
      apiGroup:
        type: string
      createdAt:
        type: string
      desc:
        type: boolean
      description:
        type: string
      id:
        type: integer
      method:
        type: string
      orderKey:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      path:
        type: string
      updatedAt:
        type: string
    type: object
...
  request.SysDictionaryDetailSearch:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      label:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      sort:
        type: integer
      status:
        type: boolean
      sysDictionaryID:
        type: integer
      updatedAt:
        type: string
      value:
        type: integer
    type: object
...
info:
  contact: { }
  description: This is a sample Server pets
  license: { }
  title: Swagger Example API
  version: 0.0.1
paths:
  /api/createApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: api路径, api中文描述, api组, 方法
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 创建基础api
      tags:
        - SysApi
  /api/deleteApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: ID
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 删除api
      tags:
        - SysApi
  /api/getAllApis:
    post:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 获取所有的Api 不分页
      tags:
        - SysApi
  /api/getApiById:
    post:
      consumes:
        - application/json
      parameters:
        - description: 根据id获取api
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.GetById'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 根据id获取api
      tags:
        - SysApi
  /api/getApiList:
    post:
      consumes:
        - application/json
      parameters:
        - description: 分页获取API列表
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.SearchApiParams'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 分页获取API列表
      tags:
        - SysApi
  /api/updateApi:
    post:
      consumes:
        - application/json
      parameters:
        - description: api路径, api中文描述, api组, 方法
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysApi'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"修改成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 创建基础api
      tags:
        - SysApi
...
  /sysDictionaryDetail/createSysDictionaryDetail:
    post:
      consumes:
        - application/json
      parameters:
        - description: SysDictionaryDetail模型
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysDictionaryDetail'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 创建SysDictionaryDetail
      tags:
        - SysDictionaryDetail
  /sysDictionaryDetail/deleteSysDictionaryDetail:
    delete:
      consumes:
        - application/json
      parameters:
        - description: SysDictionaryDetail模型
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysDictionaryDetail'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 删除SysDictionaryDetail
      tags:
        - SysDictionaryDetail
  /sysDictionaryDetail/findSysDictionaryDetail:
    get:
      consumes:
        - application/json
      parameters:
        - description: 用id查询SysDictionaryDetail
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysDictionaryDetail'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 用id查询SysDictionaryDetail
      tags:
        - SysDictionaryDetail
  /sysDictionaryDetail/getSysDictionaryDetailList:
    get:
      consumes:
        - application/json
      parameters:
        - description: 页码, 每页大小, 搜索条件
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/request.SysDictionaryDetailSearch'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 分页获取SysDictionaryDetail列表
      tags:
        - SysDictionaryDetail
  /sysDictionaryDetail/updateSysDictionaryDetail:
    put:
      consumes:
        - application/json
      parameters:
        - description: 更新SysDictionaryDetail
          in: body
          name: data
          required: true
          schema:
            $ref: '#/definitions/model.SysDictionaryDetail'
      produces:
        - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
        - ApiKeyAuth: [ ]
      summary: 更新SysDictionaryDetail
      tags:
        - SysDictionaryDetail
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"